Activate the new flocculation process
The flocculation process needed to be added to the process definition file. So, part of the changes have to do with exactly this. However, the preliminary tests showed that there was something wrong in the interfacing in flocsd.f90 to the process configuration and the actual implementation. I had to correct that and now on inspection the simple test case does look to work fine.