Correct the omission related to the DELWAQG module
For the DELWAQG module we forgot to add stoichiometry rules for organic material sinking to the bottom or being resuspended. That omission is solved here in ad hoc fashion: DetCS1 (classic) and S1_PC1 (DELWAQG) are synonymous, so I replaced the first name by the second instead of adding the missing stoichiometry rules.