UNST-4870 code clean-up and extract find_flowlink
The subroutine "cloeto1Dorbnd" returns the id of the node closest to the point [x,y]. It does this by simply looping over all flowlinks in the grid, finding the one with the shortest perpendicular distance to [x,y], and then lastly choosing between the two end nodes of that link to see which one is closest. Not very efficient, but it works.

Cleaned up that code, and extracted the bit that loops over the flowlinks to a new routine, "find_flowlink".