UNST-4870 add vicwwu to output
Added vicwwu as an output item, and renamed vicww to vicwws. vicwwu is now added to the valobs array, an output item is associated with it, and it is written to the his file.

Also removed the code that read kobs from the cache file, in compliance with Harmen's Law.