Teach DELWAQ about online hydrodynamics
Recognise the flag -onlinehydrodynamics, but only from within the BMI interface (set_var function). If the flag is passed in the DIMR configuration file, a variable internal to DELWAQ is set to prevent it from using the hydrodynamic files it normally uses. It simply ignores these files in the runtime part. As (currently) the information that is passed via BMI is limited, such quantities as the number of segments and the attributes still have to be supplied via the regular input file(s).