DELWAQ-980: Correct a mistake in handling the arguments passed via BMI
The slight refactoring of the set_var routine introduced a bug in the way the arguments were handled. This has been repaired.
Side note: the function char_array_to_string() returned a truncated result. Why that is is not clear to me, but the workaround is to do it the pedestrian way.