Build and test 'waq' for none/qp branches
This is necessary to make the testbenches and unit tests pre-merge requirements in GitHub.

- Changed qp-testbench cmake config to use waq config.
- Added none-testbench cmake config to use waq config.
- Removed branch-prefix conditions from Trigger for Test and Unit Test builds (Linux and Windows)
- Added `waq drtc` tests to none-testbench and qp-testbench columns in CSV.

After this is merged, the Test and Unit Test (for Linux and Windows) build configs can be made pre-merge requirements in GitHub.

These (as already done for 'tc' branches) tests will ensure that GitHub gets the required status report from these tests.